% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\docType{data}
\name{adduct_rules}
\alias{adduct_rules}
\title{Adduct rule table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10 rows and 3 columns.
}
\usage{
adduct_rules
}
\description{
Table with all adduct rules included by default in MetaDBparse.
}
\keyword{datasets}
