% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\docType{data}
\name{adducts}
\alias{adducts}
\title{Adduct table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 69 rows and 11 columns.
}
\usage{
adducts
}
\description{
Table with all adducts included by default in MetaDBparse.
}
\keyword{datasets}
