% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.BLOODEXPOSOME}
\alias{build.BLOODEXPOSOME}
\title{Build BLOOD EXPOSOME DB}
\usage{
build.BLOODEXPOSOME(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses the BLOOD EXPOSOME DB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.BLOODEXPOSOME(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[utils]{download.file}}
 \code{\link[openxlsx]{read.xlsx}}
 \code{\link[pbapply]{pbapply}}
 \code{\link[jsonlite]{read_json}}
}
