% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.CHEBI}
\alias{build.CHEBI}
\title{Build CHEBI}
\usage{
build.CHEBI(outfolder)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}
}
\value{
data table with parsed database
}
\description{
Parses CHEBI, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.CHEBI(outfolder=tempdir())}
}
\seealso{
\code{\link[RCurl]{getURL}}
 \code{\link[utils]{download.file}}
 \code{\link[data.table]{as.data.table}}
 \code{\link[ChemmineR]{datablock2ma}},\code{\link[ChemmineR]{datablock}}
}
