% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.DIMEDB}
\alias{build.DIMEDB}
\title{Build DIMEDB}
\usage{
build.DIMEDB(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses the DIMEDB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.DIMEDB(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[pbapply]{pbapply}}
 \code{\link[utils]{download.file}},\code{\link[utils]{unzip}}
 \code{\link[data.table]{fread}}
 \code{\link[reshape2]{cast}}
 \code{\link[Hmisc]{capitalize}}
}
