% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.DRUGBANK}
\alias{build.DRUGBANK}
\title{Build DRUGBANK}
\usage{
build.DRUGBANK(outfolder)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}
}
\value{
data table with parsed database
}
\description{
Parses the DRUGBANK DB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\details{
Requires account creation! Then please download the full XML database from the website and place in databases/drugbank_source folder. Create it if it doesn't exist yet please!
}
\seealso{
\code{\link[utils]{unzip}}
 \code{\link[stringr]{str_match}}
 \code{\link[RCurl]{getURL}}
 \code{\link[XML]{readHTMLTable}},\code{\link[XML]{xmlToList}},\code{\link[XML]{xmlValue}},\code{\link[XML]{xmlEventParse}}
 \code{\link[data.table]{as.data.table}}
 \code{\link[pbapply]{pboptions}}
}
