% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.EXPOSOMEEXPLORER}
\alias{build.EXPOSOMEEXPLORER}
\title{Build EXPOSOME EXPLORER}
\usage{
build.EXPOSOMEEXPLORER(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses the EXPOSOME EXPLORER DB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.EXPOSOMEEXPLORER(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[utils]{download.file}},\code{\link[utils]{unzip}}
 \code{\link[stringr]{str_match}}
 \code{\link[RCurl]{getURL}}
 \code{\link[data.table]{fread}}
 \code{\link[pbapply]{pbapply}}
 \code{\link[R.utils]{capitalize}}
}
