% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.FOODB}
\alias{build.FOODB}
\title{Build FOODB}
\usage{
build.FOODB(outfolder)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}
}
\value{
data table with parsed database
}
\description{
Parses the FOODB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.FOODB(outfolder=tempdir())}
}
\seealso{
\code{\link[utils]{download.file}},\code{\link[utils]{untar}}
 \code{\link[RCurl]{getURL}}
 \code{\link[stringr]{str_match}}
 \code{\link[data.table]{fread}}
}
