% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.KEGG}
\alias{build.KEGG}
\title{Build KEGG}
\usage{
build.KEGG(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses the KEGG DB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.KEGG(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[pbapply]{pbapply}}
 \code{\link[KEGGREST]{keggFind}},\code{\link[KEGGREST]{keggGet}}
 \code{\link[RCurl]{getURL}}
 \code{\link[stringr]{str_match}}
 \code{\link[data.table]{rbindlist}}
 \code{\link[rcdk]{load.molecules}},\code{\link[rcdk]{get.smiles}}
}
