% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.MASSBANK}
\alias{build.MASSBANK}
\title{Build MASSBANK DB}
\usage{
build.MASSBANK(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses MASSBANK, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.MASSBANK(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[stringr]{str_match}}
 \code{\link[utils]{download.file}},\code{\link[utils]{unzip}}
 \code{\link[pbapply]{pbapply}}
 \code{\link[data.table]{rbindlist}}
}
