% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.METABOLIGHTS}
\alias{build.METABOLIGHTS}
\title{Build METABOLIGHTS DB}
\usage{
build.METABOLIGHTS(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses the METABOLIGHTS DB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.METABOLIGHTS(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[utils]{download.file}}
 \code{\link[XML]{xmlToList}}
 \code{\link[pbapply]{pbapply}}
 \code{\link[data.table]{rbindlist}}
 \code{\link[jsonlite]{read_json}}
}
