% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.METACYC}
\alias{build.METACYC}
\title{Build METACYC}
\usage{
build.METACYC(outfolder)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}
}
\value{
data table with parsed database
}
\description{
Parses  METACYC, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\details{
Requires account creation! Then download SmartTable from 'https://trmetacyc.org/group?id=biocyc17-31223-3787684059' as 'All_compounds_of_MetaCyc.txt' and save in the databases/metacyc_source folder.
}
\seealso{
\code{\link[RCurl]{getURL}}
 \code{\link[stringr]{str_match}}
 \code{\link[data.table]{fread}}
 \code{\link[pbapply]{pbapply}}
}
