% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.mVOC}
\alias{build.mVOC}
\title{Build mVOC db}
\usage{
build.mVOC(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses the mVOC db, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.mVOC(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[XML]{getNodeSet}},\code{\link[XML]{xmlAttrs}},\code{\link[XML]{readHTMLTable}}
 \code{\link[pbapply]{pbapply}}
 \code{\link[data.table]{rbindlist}}
 \code{\link[RCurl]{getURL}}
 \code{\link[stringr]{str_match}}
}
