% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.PAMDB}
\alias{build.PAMDB}
\title{Build PAMDB}
\usage{
build.PAMDB(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses the PAMDB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.PAMDB(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[utils]{download.file}}
 \code{\link[data.table]{as.data.table}}
 \code{\link[readxl]{read_xlsx}}
 \code{\link[RCurl]{getURL}}
 \code{\link[stringr]{str_match}}
}
