% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.PHARMGKB}
\alias{build.PHARMGKB}
\title{Build PharmGKB}
\usage{
build.PHARMGKB(outfolder)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}
}
\value{
data table with parsed database
}
\description{
Parses PharmGKB drugs and chemicals (drug metabolites, etc.), returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.PHARMGKB(outfolder=tempdir())}
}
\seealso{
\code{\link[utils]{download.file}},\code{\link[utils]{unzip}}
 \code{\link[RCurl]{getURL}}
 \code{\link[stringr]{str_match}}
 \code{\link[data.table]{fread}}
}
