% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.PHENOLEXPLORER}
\alias{build.PHENOLEXPLORER}
\title{Build PHENOL EXPLORER DB}
\usage{
build.PHENOLEXPLORER(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses the PHENOL EXPLORER DB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.PHENOLEXPLORER(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[RCurl]{getURL}}
 \code{\link[stringr]{str_match}}
 \code{\link[utils]{download.file}},\code{\link[utils]{unzip}}
 \code{\link[openxlsx]{read.xlsx}}
 \code{\link[data.table]{fread}}
}
