% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.STOFF}
\alias{build.STOFF}
\title{Build STOFF db}
\usage{
build.STOFF(outfolder)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}
}
\value{
data table with parsed database
}
\description{
Parses the STOFF db, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.STOFF(outfolder=tempdir())}
}
\seealso{
\code{\link[utils]{download.file}},\code{\link[utils]{unzip}}
 \code{\link[data.table]{as.data.table}}
 \code{\link[readxl]{read_excel}}
}
