% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.VMH}
\alias{build.VMH}
\title{Build VMH}
\usage{
build.VMH(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses the VMH DB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.VMH(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[RCurl]{getURL}}
 \code{\link[stringr]{str_match}}
 \code{\link[pbapply]{pbapply}}
 \code{\link[httr]{GET}},\code{\link[httr]{content_type}},\code{\link[httr]{content}}
 \code{\link[data.table]{rbindlist}}
}
