% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.WIKIDATA}
\alias{build.WIKIDATA}
\title{Build Wikidata}
\usage{
build.WIKIDATA(outfolder)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}
}
\value{
data table with parsed database
}
\description{
Parses wikidata chemical compound database, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.WIKIDATA(outfolder=tempdir())}
}
\seealso{
\code{\link[WikidataQueryServiceR]{query_wikidata}}
 \code{\link[data.table]{as.data.table}}
}
