% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.WIKIPATHWAYS}
\alias{build.WIKIPATHWAYS}
\title{Build WikiPathways}
\usage{
build.WIKIPATHWAYS(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
data table with parsed database
}
\description{
Parses WikiPathways chemical compound database, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.WIKIPATHWAYS(outfolder=tempdir(), testMode=TRUE)}
}
\seealso{
\code{\link[SPARQL]{SPARQL}}
 \code{\link[data.table]{as.data.table}}
 \code{\link[RCurl]{getURL}}
 \code{\link[rvest]{html_nodes}}
 \code{\link[xml2]{read_html}}
}
