% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.YMDB}
\alias{build.YMDB}
\title{Build YMDB}
\usage{
build.YMDB(outfolder)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}
}
\value{
data table with parsed database
}
\description{
Parses the YMDB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.YMDB(outfolder=tempdir())}
}
\seealso{
\code{\link[utils]{download.file}},\code{\link[utils]{unzip}}
 \code{\link[data.table]{as.data.table}},\code{\link[data.table]{fread}},\code{\link[data.table]{rbindlist}}
 \code{\link[ChemmineR]{datablock2ma}},\code{\link[ChemmineR]{datablock}}
 \code{\link[pbapply]{pbapply}}
 \code{\link[RCurl]{getURL}}
 \code{\link[stringr]{str_match}}
}
