% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-from_formula.R
\name{chemspiderInfo}
\alias{chemspiderInfo}
\title{Find more info through ChemSpider.}
\usage{
chemspiderInfo(ids, maxn = 100, apikey)
}
\arguments{
\item{ids}{Character vector of ChemSpider IDs.}

\item{maxn}{Max ids per batch (batch search is used), Default: 100}

\item{apikey}{ChemSpider API key}
}
\value{
Data table with match results
}
\description{
Takes ChemSpider CIDs and finds name, SMILES, citations on pubmed/references.
}
\seealso{
\code{\link[gsubfn]{fn}}
 \code{\link[httr]{POST}},\code{\link[httr]{add_headers}}
 \code{\link[jsonlite]{read_json}}
}
