% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_extended.R
\name{doIsotopes}
\alias{doIsotopes}
\title{Generate isotopes for given formula}
\usage{
doIsotopes(formula, charge, count.isos = F)
}
\arguments{
\item{formula}{Molecular formula}

\item{charge}{Final charge}

\item{count.isos}{Add columns for amounts of 2H, 13C, 15N atoms? Useful for heavy isotope experiments.}
}
\value{
Table with isotopes of this molecular formula
}
\description{
Takes in formula and returns isotope pattern m/z values.
}
\examples{
 doIsotopes(formula="C6H12O6", charge=0)
}
\seealso{
\code{\link[enviPat]{isopattern}}
}
