% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-biotransformer.R
\name{downloadBT}
\alias{downloadBT}
\title{Download biotransformer jar file.}
\usage{
downloadBT(outfolder)
}
\arguments{
\item{outfolder}{In which folder do you want to save the .jar file?}
}
\value{
File location
}
\description{
To predict enzymatic metabolite changes, you can enable Biotransformer processing in the base database creation. This downloads the necessary JAR file from the developer page.
}
\examples{
 \dontrun{jarloc = downloadBT(outfolder = tempdir())}
}
\seealso{
\code{\link[utils]{download.file}},\code{\link[utils]{unzip}}
}
