% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-from_formula.R
\name{filterFormula}
\alias{filterFormula}
\title{Apply seven golden rules to a vector of formulas}
\usage{
filterFormula(formulas, rules = c("senior", "lewis", "hc", "chnops", "nops"))
}
\arguments{
\item{formulas}{Molecular formulas}

\item{rules}{Rules to apply. Default: c("senior", "lewis", "hc", "chnops", "nops")}
}
\value{
Vector of formulas that pass rules
}
\description{
Returns formulas that pass the user-given rules.
}
\examples{
filterFormula("C6H12O6")
}
\seealso{
\code{\link[enviPat]{check_chemform}}
 \code{\link[data.table]{rbindlist}}
 \code{\link[stringr]{str_match}}
}
