% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_base.R
\name{iatom.to.charge}
\alias{iatom.to.charge}
\title{Get formal charge from iatomcontainer}
\usage{
iatom.to.charge(iatoms, silent = TRUE)
}
\arguments{
\item{iatoms}{list of rcdk iatomcontainers}

\item{silent}{suppress warnings?, Default: TRUE}
}
\value{
Character vector of formal charges per iatomcontainer.
}
\description{
This function takes iatomcontainer object and returns the formal charge.
}
\examples{
 iatom = smiles.to.iatom(c('OC[C@H]1OC(O)[C@H](O)[C@H](O)[C@H]1O'))
 ch = iatom.to.charge(iatom)
}
\seealso{
\code{\link[rcdk]{get.total.formal.charge}}
 \code{\link[rJava]{jcall}}
}
