% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_base.R
\name{iatom.to.formula}
\alias{iatom.to.formula}
\title{Get molecular formula from iatomcontainer}
\usage{
iatom.to.formula(iatoms, silent = TRUE)
}
\arguments{
\item{iatoms}{list of rcdk iatomcontainers}

\item{silent}{suppress warnings?, Default: TRUE}
}
\value{
Character vector of formulas per iatomcontainer.
}
\description{
This function takes iatomcontainer object and returns the molecular formula.
}
\examples{
 iatom = smiles.to.iatom(c('OC[C@H]1OC(O)[C@H](O)[C@H](O)[C@H]1O'))
 form = iatom.to.formula(iatom)
}
\seealso{
\code{\link[rcdk]{get.mol2formula}}
 \code{\link[rJava]{jcall}}
}
