% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_base.R
\name{iatom.to.smiles}
\alias{iatom.to.smiles}
\title{Get SMILES from iatom container}
\usage{
iatom.to.smiles(iatoms, smitype = "Canonical", silent = TRUE)
}
\arguments{
\item{iatoms}{list of iatomcontainers}

\item{smitype}{Which type of SMILES to export?, Default: 'Canonical'}

\item{silent}{Suppress warnings?, Default: TRUE}
}
\value{
character vector of SMILES in the chosen format
}
\description{
This function takes an rcdk iatomcontainer and returns SMILES
}
\examples{
 iatom = smiles.to.iatom(c('OC[C@H]1OC(O)[C@H](O)[C@H](O)[C@H]1O'))
 smi = iatom.to.smiles(iatom)
}
\seealso{
\code{\link[rcdk]{get.smiles}},\code{\link[rcdk]{smiles.flavors}}
 \code{\link[rJava]{jcall}}
}
