% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_extended.R
\docType{data}
\name{isotopes}
\alias{isotopes}
\title{Weights of all isotopes of all elemnts used}
\format{
An object of class \code{data.frame} with 308 rows and 5 columns.
}
\usage{
isotopes
}
\description{
Sourced from 'enviPat' package
}
\keyword{datasets}
