% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\docType{data}
\name{lmdb}
\alias{lmdb}
\title{LMDB}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1070 rows and 6 columns.
}
\usage{
lmdb
}
\description{
LMDB database, included with permission from the DB creators.
}
\keyword{datasets}
