% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_basics.R
\name{openBaseDB}
\alias{openBaseDB}
\title{Create/open and prepare SQLite database}
\usage{
openBaseDB(outfolder, dbname)
}
\arguments{
\item{outfolder}{Which folder are you building your databases in?}

\item{dbname}{What is the name of the database? (exclude .db)}
}
\value{
Nothing, writes SQLITE database to outfolder
}
\description{
Create/open and prepare SQLite database
}
\seealso{
\code{\link[RSQLite]{SQLite}}
 \code{\link[DBI]{dbExecute}}
}
