% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-from_formula.R
\name{pubChemInfo}
\alias{pubChemInfo}
\title{Find additional info on a PubChem ID.}
\usage{
pubChemInfo(ids, maxn = 30)
}
\arguments{
\item{ids}{Vector of identifiers.}

\item{maxn}{Compounds searched per batch search, Default: 30}
}
\value{
Table with additional info on PubChem IDs
}
\description{
Takes PubChem ID and finds name, formula, smiles, charge
}
\examples{
pubChemInfo(c(5793))
}
\seealso{
\code{\link[data.table]{rbindlist}}
}
