% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_basics.R
\name{removeDB}
\alias{removeDB}
\title{Remove database}
\usage{
removeDB(outfolder, dbname)
}
\arguments{
\item{outfolder}{Which folder are you building your databases in?}

\item{dbname}{What database do you want to remove? (exclude .db suffix)}
}
\value{
Nothing, removes database from disk
}
\description{
Removes database from disk completely.
}
