% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_extended.R
\name{removeFailedStructures}
\alias{removeFailedStructures}
\title{Remove structures where isotope generation failed.}
\usage{
removeFailedStructures(outfolder, ext.dbname = "extended")
}
\arguments{
\item{outfolder}{Which folder are your databases in?}

\item{ext.dbname}{Extended database name (without .db suffix), Default: 'extended'}
}
\description{
Sometimes if a run crashes, or a structure is bugged somehow, it is still registered as 'done' in the extended database and cannot be redone. This function removes these structures. Warning: slow!
}
\seealso{
\code{\link[RSQLite]{SQLite}}
}
