% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-from_formula.R
\name{revertAdduct}
\alias{revertAdduct}
\title{Break formula apart into adduct and main formula}
\usage{
revertAdduct(formula, add_name, adduct_table = adducts)
}
\arguments{
\item{formula}{Formula of interest}

\item{add_name}{Adduct names to consider ('Name' column of adduct table)}

\item{adduct_table}{Full adduct table, Default: adducts}
}
\value{
Table with formula, adduct, isotope
}
\description{
Used to use the formula creation function and consider adducts at the same time.
}
\examples{
revertAdduct("C6H13O6", "[M+H]1+")
}
\seealso{
\code{\link[enviPat]{check_ded}},\code{\link[enviPat]{subform}},\code{\link[enviPat]{mergeform}},\code{\link[enviPat]{multiform}}
}
