% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_basics.R
\name{sdfStream.joanna}
\alias{sdfStream.joanna}
\title{Adjusted sdfStream version for databases that store their compounds in SDF format}
\usage{
sdfStream.joanna(
  input,
  output,
  append = FALSE,
  fct,
  Nlines = 10000,
  startline = 1,
  restartNlines = 10000,
  silent = FALSE,
  ...
)
}
\arguments{
\item{input}{input SDF file}

\item{output}{output CSV file to write to}

\item{append}{Append to existing CSV file or start anew?, Default: FALSE}

\item{fct}{Function to apply to each object to get the wanted columns}

\item{Nlines}{Lines to read in one go?, Default: 10000}

\item{startline}{Start at line, Default: 1}

\item{restartNlines}{Restart after x lines, Default: 10000}

\item{silent}{Suppress warnings?, Default: FALSE}

\item{...}{Other arguments}
}
\value{
Nothing, writes csv version of given SDF files to disk
}
\description{
Adjusted from existing function to extract the columns needed for MetaDBparse database format
}
