% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_search.R
\name{showAllBase}
\alias{showAllBase}
\title{Show all compounds in base db}
\usage{
showAllBase(outfolder, base.dbname)
}
\arguments{
\item{outfolder}{Which folder is your database in?}

\item{base.dbname}{Which base database do you want to explore? (exclude .db suffix)}
}
\value{
Data table with whole database.
}
\description{
Shows all compounds in this base database as data table.
}
\details{
This may be quite memory consuming for larger databases!!
}
\examples{
 \dontrun{myFolder = tempdir()}
 \dontrun{buildBaseDB(outfolder = myFolder, "lmdb", test = TRUE)}
 \dontrun{showAllBase(outfolder = myFolder, "lmdb")}
}
\seealso{
\code{\link[RSQLite]{SQLite}}
}
