% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_basics.R
\name{writeDB}
\alias{writeDB}
\title{Write table to SQLite database}
\usage{
writeDB(conn, table, tblname)
}
\arguments{
\item{conn}{Database connection (from DBI::dbConnect or similar)}

\item{table}{Data frame or data table}

\item{tblname}{SQLITE table name to append to}
}
\value{
Nothing, writes table to SQLITE database
}
\description{
Simple wrapper - take data table or data frame and append it to the given table in the given database.
}
\seealso{
\code{\link[DBI]{dbWriteTable}}
}
