% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oxy-db_parsers.R
\name{build.RMDB}
\alias{build.RMDB}
\title{Build RMDB (deprecated)}
\usage{
build.RMDB(outfolder, testMode = FALSE)
}
\arguments{
\item{outfolder}{Which folder to save temp files to?}

\item{testMode}{run in test mode? Only parses first ten compounds}
}
\value{
Message that RMDB is deprecated
}
\description{
Parses the RMDB, returns data table with columns compoundname, description, charge, formula and structure (in SMILES)
}
\examples{
\dontrun{build.RMDB(outfolder=tempdir(), testMode=TRUE)}
}
